Move hands related functions to minimalin.c
